Autosubmit: add WRONG_LEVEL response
It seems that this year the form does not include the "level" parameter, so the site expects you to always submit the current part of the puzzle (in other words, if you've solved the first part of the puzzle, you cannot submit it again).

The WRONG_LEVEL result will not be stored in the results cache.